app/back/requirements.txt
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
face-recognition==1.3.0
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
faiss-cpu==1.7.4
pandas==2.0.2
